% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{parametric}
\alias{parametric}
\title{Fully Parametric Survival Model}
\usage{
parametric(
  formula = NULL,
  data = NULL,
  reverse = FALSE,
  time_variable = "time",
  status_variable = "status",
  x = NULL,
  y = NULL,
  eps = 1e-15,
  ...
)
}
\arguments{
\item{formula}{\code{(formula(1))}\cr
Object specifying the model fit, left-hand-side of formula should describe a \code{\link[survival:Surv]{survival::Surv()}}
object.}

\item{data}{\code{(data.frame(1))}\cr
Training data of \code{data.frame} like object, internally is coerced with \code{\link[stats:model.matrix]{stats::model.matrix()}}.}

\item{reverse}{\code{(logical(1))}\cr
If \code{TRUE} fits estimator on censoring distribution, otherwise (default) survival distribution.}

\item{time_variable}{\code{(character(1))}\cr
Alternative method to call the function. Name of the 'time' variable, required if \code{formula}.
or \code{x} and \code{Y} not given.}

\item{status_variable}{\code{(character(1))}\cr
Alternative method to call the function. Name of the 'status' variable, required if \code{formula}
or \code{x} and \code{Y} not given.}

\item{x}{\code{(data.frame(1))}\cr
Alternative method to call the function. Required if \verb{formula, time_variable} and
\code{status_variable} not given. Data frame like object of features which is internally
coerced with \code{model.matrix}.}

\item{y}{\verb{([survival::Surv()])}\cr
Alternative method to call the function. Required if \verb{formula, time_variable} and
\code{status_variable} not given. Survival outcome of right-censored observations.}

\item{eps}{\code{(numeric(1))} \cr
Used when the fitted \code{scale} parameter is too small. Default \code{1e-15}.}

\item{...}{\code{ANY} \cr
Additional arguments passed to \code{\link[survival:survreg]{survival::survreg()}}.}
}
\value{
An object inheriting from class \code{parametric}.
}
\description{
Fit/predict implementation of \code{\link[survival:survreg]{survival::survreg()}}, which can return
absolutely continuous distribution predictions using \pkg{distr6}.
}
\examples{
if (requireNamespaces(c("distr6", "survival"))) {
 library(survival)
 parametric(Surv(time, status) ~ ., data = simsurvdata(10))
}
}
