% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtime.R
\name{predict.coxtime}
\alias{predict.coxtime}
\title{Predict Method for Cox-Time Survival Neural Network}
\usage{
\method{predict}{coxtime}(
  object,
  newdata,
  batch_size = 256L,
  num_workers = 0L,
  type = c("survival", "risk", "all"),
  distr6 = FALSE,
  ...
)
}
\arguments{
\item{object}{(\code{akritascoxtime})\cr
Object of class inheriting from \code{"coxtime"}.}

\item{newdata}{\code{(data.frame(1))}\cr
Testing data of \code{data.frame} like object, internally is coerced with \code{\link[stats:model.matrix]{stats::model.matrix()}}.
If missing then training data from fitted object is used.}

\item{batch_size}{\code{(integer(1))}\cr
Passed to \code{pycox.models.Coxtime.fit}, elements in each batch.}

\item{num_workers}{\code{(integer(1))}\cr
Passed to \code{pycox.models.Coxtime.fit}, number of workers used in the dataloader.}

\item{type}{(\code{numeric(1)})\cr
Type of predicted value. Choices are survival probabilities over all time-points in training
data (\code{"survival"}) or a relative risk ranking (\code{"risk"}), which is the mean cumulative hazard
function over all time-points, or both (\code{"all"}).}

\item{distr6}{\code{(logical(1))}\cr
If \code{FALSE} (default) and \code{type} is \code{"survival"} or \code{"all"} returns data.frame of survival
probabilities, otherwise returns a \code{\link[distr6:VectorDistribution]{distr6::VectorDistribution()}}.}

\item{...}{\code{ANY} \cr
Currently ignored.}
}
\value{
An object inheriting from class \code{coxtime}.
}
\description{
Predicted values from a fitted Cox-Time ANN.
}
\references{
Kvamme, H., Borgan, Ø., & Scheel, I. (2019).
Time-to-event prediction with neural networks and Cox regression.
Journal of Machine Learning Research, 20(129), 1–30.
}
