% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{predict.parametric}
\alias{predict.parametric}
\title{Predict method for Parametric Model}
\usage{
\method{predict}{parametric}(
  object,
  newdata,
  form = c("aft", "ph", "tobit", "po"),
  times = NULL,
  return_method = c("continuous", "discrete"),
  type = c("survival", "risk", "all"),
  distr6 = FALSE,
  ntime = 150,
  round_time = 2,
  ...
)
}
\arguments{
\item{object}{(\code{parametric(1)})\cr
Object of class inheriting from \code{"parametric"}.}

\item{newdata}{\code{(data.frame(1))}\cr
Testing data of \code{data.frame} like object, internally is coerced with \code{\link[stats:model.matrix]{stats::model.matrix()}}.
If missing then training data from fitted object is used.}

\item{form}{\code{(character(1))} \cr
The form of the predicted distribution, see \code{details} for options.}

\item{times}{\code{(numeric())}\cr
Times at which to evaluate the estimator. If \code{NULL} (default) then evaluated at all unique times
in the training set.}

\item{type}{(\code{character(1)})\cr
Type of predicted value. Choices are survival probabilities over all time-points in training
data (\code{"survival"}) or a relative risk ranking (\code{"risk"}), which is the sum of the predicted
cumulative hazard function so higher rank implies higher risk of event, or both (\code{"all"}).}

\item{distr6}{(\code{logical(1)})\cr
If \code{FALSE} (default) and \code{type} is \code{"survival"} or \code{"all"} returns matrix of survival
probabilities, otherwise returns a \code{\link[distr6:Distribution]{distr6::Distribution()}}.}

\item{ntime}{\code{(numeric(1))}\cr
Number of unique time-points in the training set, default is 150.}

\item{round_time}{\code{(numeric(1))}\cr
Number of decimal places to round time-points to, default is 2, set to \code{FALSE} for no rounding.}

\item{...}{\code{ANY} \cr
Currently ignored.}
}
\value{
A \code{numeric} if \code{type = "risk"}, a \code{\link[distr6:Matdist]{distr6::Matdist()}} (if \code{distr6 = TRUE})
and \code{type = "survival"}; a \code{matrix} if (\code{distr6 = FALSE}) and \code{type = "survival"} where
entries are survival probabilities with rows of observations and columns are time-points;
or a list combining above if \code{type = "all"}.
}
\description{
Predicted values from a fitted Parametric survival model.
}
\details{
The \code{form} parameter determines how the distribution is created.
Options are:
\itemize{
\item Accelerated failure time (\code{"aft"}) \deqn{h(t) = h_0(\frac{t}{exp(lp)})exp(-lp)}
\item Proportional Hazards (\code{"ph"}) \deqn{h(t) = h_0(t)exp(lp)}
\item Tobit (\code{"tobit"}) \deqn{h(t) = \Phi(\frac{t - lp}{scale})}
\item Proportional odds (\code{"po"}) \deqn{h(t) = \frac{h_0(t)}{1 + (exp(lp)-1)S_0(t)}}
}

where \eqn{h_0,S_0} are the estimated baseline hazard and survival functions
(in this case with a given parametric form), \eqn{lp} is the predicted linear
predictor calculated using the formula \eqn{lp = \hat{\beta} X_{new}} where
\eqn{X_{new}} are the variables in the test data set and \eqn{\hat{\beta}}
are the coefficients from the fitted parametric survival model (\code{object}).
\eqn{\Phi} is the cdf of a N(0, 1) distribution, and \eqn{scale} is the
fitted scale parameter.
}
\examples{
if (requireNamespaces(c("distr6", "survival"))) {
  library(survival)

  set.seed(42)
  train <- simsurvdata(50)
  test <- simsurvdata(50)
  fit <- parametric(Surv(time, status) ~ ., data = train)
  predict(fit, newdata = test)

  # distr6 parameter ignored when returning continuous distribution
  predict_distr <- predict(fit, newdata = test, distr6 = FALSE)
  predict_distr$survival(3)

  # Return a discrete distribution survival matrix
  predict_distr <- predict(fit, newdata = test,
    return_method = "discrete", distr6 = FALSE)
  predict_distr[1:5, 1:5]

  # Return a relative risk ranking with type = "risk"
  predict(fit, newdata = test, type = "risk")[1:5]

  # Or survival probabilities and a rank
  predict(fit, newdata = test, type = "all", distr6 = TRUE)
}
}
