% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{requireNamespaces}
\alias{requireNamespaces}
\title{Vectorised Logical requireNamespace}
\usage{
requireNamespaces(x)
}
\arguments{
\item{x}{\code{(character())} \cr string naming the packages/name spaces to load.}
}
\description{
Vectorises the \link{requireNamespace} function and returns \code{TRUE} if all packages, \code{x},
are available and \code{FALSE} otherwise.
}
\examples{
requireNamespaces("survivalmodels")
requireNamespaces(c("survivalmodels", "distr6"))

# test for some fake package
requireNamespaces(c("survivalmodels", "distr6", "111"))

}
