% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsurvdata.R
\name{simsurvdata}
\alias{simsurvdata}
\title{Simulate Survival Data}
\usage{
simsurvdata(n = 100, trt = 2, age = 2, sex = 1.5, cutoff = NULL, cens = 0.3)
}
\arguments{
\item{n}{\code{(integer(1))} \cr Number of samples}

\item{trt, age, sex}{\code{(numeric(1))} \cr Coefficients for covariates.}

\item{cutoff}{\code{(numeric(1))} \cr Deprecated, in future use \code{cens}.}

\item{cens}{\code{(numeric(1))} \cr Proportion of censoring to be generated, cut-off time is then
selected as the quantile that results in \code{cens}.}
}
\value{
\code{\link[=data.frame]{data.frame()}}
}
\description{
Function for simulating survival data.
}
\details{
Currently limited to three covariates, Weibull survival times, and Type I censoring.
This will be expanded to a flexible simulation function in future updates. For now the function
is primarily limited to helping function examples.
}
\examples{
simsurvdata()

}
