% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalmodels-package.R
\docType{package}
\name{survivalmodels-package}
\alias{survivalmodels}
\alias{survivalmodels-package}
\title{survivalmodels: Models for Survival Analysis}
\description{
survivalmodels implements classical and machine learning models for survival analysis that either
do not already exist in R or for more efficient implementations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RaphaelS1/survivalmodels/}
  \item Report bugs at \url{https://github.com/RaphaelS1/survivalmodels/issues}
}

}
\author{
\strong{Maintainer}: Yohann Foucher \email{yohann.foucher@univ-poitiers.fr} (\href{https://orcid.org/0000-0003-0330-7457}{ORCID})

Authors:
\itemize{
  \item Raphael Sonabend (\href{https://orcid.org/0000-0001-9225-4654}{ORCID})
}

}
