% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{predict.parametric}
\alias{predict.parametric}
\title{Predict method for Parametric Model}
\usage{
\method{predict}{parametric}(
  object,
  newdata,
  form = c("aft", "ph", "tobit", "po"),
  times = NULL,
  return_method = c("continuous", "discrete"),
  type = c("survival", "risk", "all"),
  distr6 = FALSE,
  ntime = 150,
  round_time = 2,
  ...
)
}
\arguments{
\item{object}{(\code{parametric(1)})\cr
Object of class inheriting from \code{"parametric"}.}

\item{newdata}{\code{(data.frame(1))}\cr
Testing data of \code{data.frame} like object, internally is coerced with \code{\link[stats:model.matrix]{stats::model.matrix()}}.
If missing then training data from fitted object is used.}

\item{form}{\code{(character(1))} \cr
The form of the predicted distribution, see \code{details} for options.}

\item{times}{\code{(numeric())}\cr
Times at which to evaluate the estimator. If \code{NULL} (default) then evaluated at all unique times
in the training set.}

\item{return_method}{(\code{character(1)}) \cr
If \code{"continuous"} (default) then distribution returned as a continuous
distribution using \pkg{distr6} and the \code{distr6} parameter is ignored in this case.
Otherwise if \code{"discrete"} then distribution either returned as a matrix or discrete
distribution depending on \code{distr6} parameter.}

\item{type}{(\code{character(1)})\cr
Type of predicted value. Choices are survival probabilities over all time-points in training
data (\code{"survival"}) or a relative risk ranking (\code{"risk"}), which is the sum of the predicted
cumulative hazard function so higher rank implies higher risk of event, or both (\code{"all"}).}

\item{distr6}{(\code{logical(1)})\cr
If \code{FALSE} (default) and \code{type} is \code{"survival"} or \code{"all"} returns matrix of survival
probabilities, otherwise returns a \code{\link[distr6:Matdist]{distr6::Matdist()}}.}

\item{ntime}{\code{(numeric(1))}\cr
Number of unique time-points in the training set, default is 150.}

\item{round_time}{\code{(numeric(1))}\cr
Number of decimal places to round time-points to, default is 2, set to \code{FALSE} for no rounding.}

\item{...}{\code{ANY} \cr
Currently ignored.}
}
\value{
A \code{numeric} if \code{type = "risk"}, a \code{\link[distr6:Matdist]{distr6::Matdist()}} (if \code{distr6 = TRUE})
and \code{type = "survival"}; a \code{matrix} if (\code{distr6 = FALSE}) and \code{type = "survival"} where
entries are survival probabilities with rows of observations and columns are time-points;
or a list combining above if \code{type = "all"}.
}
\description{
Predicted values from a fitted Parametric survival model.
}
\details{
The \code{form} parameter determines how the distribution is created.
Options are:
\itemize{
\item Accelerated failure time (\code{"aft"}) - $h(t) = exp(-lp)h0(t/exp(lp))$
\item Proportional Hazards (\code{"ph"}) - $h(t) = h0(t)exp(lp)$
\item Tobit (\code{"tobit"}) - Φ((t - lp) / scale)
\item Proportional odds (\code{"po"}) - $h(t) = /h0(t){1 + (exp(lp)-1)S0(t)}^-1$
}
}
\examples{
if (requireNamespaces(c("distr6", "survival"))) {

library(survival)

train <- 1:10
test <- 11:20
fit <- parametric(Surv(time, status) ~ ., data = rats[train, ])
predict(fit, newdata = rats[test, ])

# Use distr6 = TRUE to return a distribution
predict_distr <- predict(fit, newdata = rats[test, ], distr6 = TRUE)
predict_distr$survival(100)

# Return a relative risk ranking with type = "risk"
predict(fit, newdata = rats[test, ], type = "risk")

# Or survival probabilities and a rank
predict(fit, newdata = rats[test, ], type = "all", distr6 = TRUE)
}
}
