% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_seed}
\alias{set_seed}
\title{Set seed in R numpy and torch}
\usage{
set_seed(seed_R, seed_np = seed_R, seed_torch = seed_R)
}
\arguments{
\item{seed_R}{(\code{integer(1)}) \cr \code{seed} passed to \link{set.seed}.}

\item{seed_np}{(\code{integer(1)}) \cr \code{seed} passed to \code{numpy$random$seed}. Default is same as \code{seed_R}.}

\item{seed_torch}{(\code{integer(1)}) \cr \code{seed} passed to \code{numpy$random$seed}.
Default is same as \code{seed_R}.}
}
\value{
No return value.
}
\description{
To ensure consistent results, a seed has to be set in R
using \code{set.seed} as usual but also in  \code{numpy} and  \code{torch} via  \code{reticulate}.
Therefore this function simplifies the process into one funciton.
}
